+++ test/unit/config.test.js
@@ -1,4 +1,5 @@
 
+jest.mock('../../src/plugins/worker')
 import nexteinConfig from '../../src/config'
 
 describe('config', () => {